requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
flasgger==0.9.7.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-dotenv==1.0.1
gunicorn==23.0.0
